style(desktop): icons for changelog/discord rows; modal-native keyboard shortcuts dialog
