+++ apps/packages/ui/src/icons.tsx
@@ -4,6 +4,6 @@ export type { IconProps } from "./icons/types";
 export { ChevronRight, ChevronDown, ChevronUpDown, Minus, Plus, MiniPlus, X, Copy, Search, Hash, Keyboard, KeyRound, Pencil, Trash, Archive, Clock, Settings, SlidersHorizontal, Palette, CircleQuestion, LifeBuoy, ArrowLeft, ArrowRight, LogOut, ArrowUp, ListFilter, StopSquare, Play, Pause, ArrowUpRight, ExternalLink, Check, XLines, Fork, Undo, Link2, WrapText, MoreHorizontal, SendIcon } from "./icons/core";
 export { GitMerge, GitBranch, GitPullRequest, Terminal, SquareTerminal, FileText, FileIcon, InlinePathMentionIcon, FileCode, StackedFiles, FilePen, FilePlus, FolderPlus, FolderList, Folder, FolderOpen, FolderClosedFilled, FolderPlusFilled, FolderFilled, Filter, CollapseAll, ExpandAll, GitCommit, Tree, GitBranchIcon, ProjectNotebook } from "./icons/workspace";
 export { Robot, ClipboardList, ClipboardListFilled, ReadModeFilled, EditModeFilled, BuildModeFilled, OpencodeBuildModeFilled, OpencodePlanModeFilled, MessageSquareFilled, AddMessage, AddPlan, Brain, ThinkingGlyph, PlanningIcon, Shield, MessageSquare, StickyNote, ScratchPadIcon, Zap, Sparkles, ClaudeSparkle, BotMessageSquare } from "./icons/product";
-export { Mail, Calendar, CalendarClock, CreditCard, GitHub, Globe, Smartphone, Sun, Moon, Monitor, Home, Grid, CloudUpload, Blocks, CircleUser, Building2, UsersRound, CloudIcon, Server, RefreshCw } from "./icons/platform";
+export { Mail, Calendar, CalendarClock, CreditCard, Discord, GitHub, Globe, Smartphone, Sun, Moon, Monitor, Home, Grid, CloudUpload, Blocks, CircleUser, Building2, UsersRound, CloudIcon, Server, RefreshCw } from "./icons/platform";
 export { CheckCircleFilled, Circle, CircleAlert, ShieldCheckFilled, Spinner, BrailleSweepBadge } from "./icons/status";
 export { AppShellReviewIcon, AppShellBrowserIcon, AppShellTerminalIcon, AppShellPlusIcon, AppShellPanelToggleIcon, AppShellTabCloseIcon, SplitPanel, SplitPanelRight } from "./icons/app-shell";
